+++ react/Shelf.js
@@ -25,11 +25,12 @@ const { ProductListProvider } = ProductListContext
 /**
  * Shelf Component. Queries a list of products and shows them.
  */
-const Shelf = ({
-  data,
-  productList = ProductList.defaultProps,
-  paginationDotsVisibility = 'visible',
-}) => {
+const Shelf = props => {
+  const {
+    data,
+    paginationDotsVisibility = 'visible',
+    productList = ProductList.defaultProps,
+  } = props
   const handles = useCssHandles(CSS_HANDLES)
   const { isMobile } = useDevice()
   const { loading, error, products } = data || {}
@@ -38,16 +39,13 @@ const Shelf = ({
     return products && products.map(normalizeBuyable).filter(Boolean)
   }, [products])
 
-  const productListProps = useMemo(
-    () => ({
-      products: filteredProducts,
-      loading: loading,
-      isMobile,
-      paginationDotsVisibility,
-      ...productList,
-    }),
-    [filteredProducts, loading, isMobile, paginationDotsVisibility, productList]
-  )
+  const productListProps = {
+    ...productList,
+    isMobile,
+    loading: loading,
+    paginationDotsVisibility,
+    products: filteredProducts,
+  }
 
   if (loading) {
     return <Loading />
@@ -120,70 +118,8 @@ const options = {
 
 const EnhancedShelf = graphql(productsQuery, options)(Shelf)
 
-EnhancedShelf.getSchema = props => {
-  return {
-    title: 'admin/editor.shelf.title',
-    description: 'admin/editor.shelf.description',
-    type: 'object',
-    properties: {
-      category: {
-        title: 'admin/editor.shelf.category.title',
-        description: 'admin/editor.shelf.category.description',
-        type: 'string',
-        isLayout: false,
-      },
-      specificationFilters: {
-        title: 'admin/editor.shelf.specificationFilters.title',
-        type: 'array',
-        items: {
-          title: 'admin/editor.shelf.specificationFilters.item.title',
-          type: 'object',
-          properties: {
-            id: {
-              type: 'string',
-              title: 'admin/editor.shelf.specificationFilters.item.id.title',
-            },
-            value: {
-              type: 'string',
-              title: 'admin/editor.shelf.specificationFilters.item.value.title',
-            },
-          },
-        },
-      },
-      collection: {
-        title: 'admin/editor.shelf.collection.title',
-        type: 'string',
-        isLayout: false,
-      },
-      orderBy: {
-        title: 'admin/editor.shelf.orderBy.title',
-        type: 'string',
-        enum: getOrdenationValues(),
-        enumNames: getOrdenationNames(),
-        default: OrdenationTypes.ORDER_BY_RELEVANCE.value,
-        isLayout: false,
-      },
-      productList: ProductList.getSchema(props),
-      hideUnavailableItems: {
-        title: 'admin/editor.shelf.hideUnavailableItems',
-        type: 'boolean',
-        default: false,
-        isLayout: false,
-      },
-      skusFilter: {
-        title: 'admin/editor.shelf.skusFilter',
-        description: 'admin/editor.shelf.skusFilter.description',
-        type: 'string',
-        default: 'ALL_AVAILABLE',
-        enum: ['ALL_AVAILABLE', 'ALL', 'FIRST_AVAILABLE'],
-        enumNames: [
-          'admin/editor.shelf.skusFilter.all-available',
-          'admin/editor.shelf.skusFilter.none',
-          'admin/editor.shelf.skusFilter.first-available',
-        ],
-      },
-    },
-  }
+EnhancedShelf.schema = {
+  title: 'admin/editor.shelf.title',
 }
 
 export default EnhancedShelf